backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
openai
yfinance
python-dotenv
aiosqlite
pytest
pytest-asyncio
